requirements.txt
fastapi
uvicorn
google-generativeai
requests
python-dotenv
pydantic-settings
pydantic
